Require all actions and guards to fptrs compatible
With this change the input expression can be discarded for statemachine
init. Instead only the type of the input expression is used. So besides
the fact that the tables have to be made available (and are currently
function local statics), the initialization can be expressed at compile
time.

The user facing change is that actions and guards have to be stateless
and shall not reference anything. Instead the functions get access to a
context object that the user has to provide via start and
process_event.